\name{setops}
\alias{intersect}
%  \alias{setdiff}
%\alias{setequal}
\alias{setops}
\alias{union}
\title{Set operations.}
\usage{
intersect(x, y, ...)

union(x, y, ...)

%setdiff(x, y, ...)

%setequal(x, y, ...)
}
\arguments{
\item{x,y}{objects to compare (ignoring order)}

\item{...}{other arguments passed on to methods}
}
\description{
Support for these operations is experimental and limited to \code{intersect} and \code{union}. For the generic function documentationSee \code{\link[dplyr]{setops}}
}

